requirements.txt
python-nmap
sockets